refactor: extract secrets routes from router
Move repo and global secret handlers into a dedicated routes module and share repository resolution logic across route modules to keep router.ts focused on route composition.
